pyproject.toml
[project]
name = "cvtools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyaml>=26.2.1",
]
